# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SwarmSummaryStateBreakdown = {
  queued: number;
  running: number;
  completed: number;
  failed: number;
};

export interface TrendPoint {
  label: string;
  value: number;
}

export interface SwarmSummary {
  activeObjectives: number;
  activeAgents: number;
  successRate: number;
  averageFitness: number;
  totalSpawns: number;
  mutationPressure: number;
  stateBreakdown: SwarmSummaryStateBreakdown;
  fitnessTrend: TrendPoint[];
  updatedAt: string;
}

export type ObjectiveStatus = typeof ObjectiveStatus[keyof typeof ObjectiveStatus];


export const ObjectiveStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export type ObjectivePriority = typeof ObjectivePriority[keyof typeof ObjectivePriority];


export const ObjectivePriority = {
  low: 'low',
  normal: 'normal',
  high: 'high',
  critical: 'critical',
} as const;

export interface Objective {
  id: string;
  title: string;
  description: string;
  status: ObjectiveStatus;
  priority: ObjectivePriority;
  progress: number;
  agentCount: number;
  /** @nullable */
  rootAgentId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ObjectiveInputPriority = typeof ObjectiveInputPriority[keyof typeof ObjectiveInputPriority];


export const ObjectiveInputPriority = {
  low: 'low',
  normal: 'normal',
  high: 'high',
  critical: 'critical',
} as const;

export interface ObjectiveInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  priority: ObjectiveInputPriority;
}

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Agent {
  id: string;
  name: string;
  status: AgentStatus;
  strategy: string;
  frame: string;
  fitness: number;
  objectiveId: string;
  /** @nullable */
  parentId?: string | null;
  generation: number;
  runtimeMs: number;
  state: string;
  /** @nullable */
  failureSignature?: string | null;
  startedAt: string;
}

export type GenotypeTrend = typeof GenotypeTrend[keyof typeof GenotypeTrend];


export const GenotypeTrend = {
  up: 'up',
  down: 'down',
  flat: 'flat',
} as const;

export interface Genotype {
  id: string;
  frame: string;
  label: string;
  strategy: string;
  adjustedScore: number;
  fertilityScore: number;
  totalSpawns: number;
  successRate: number;
  cumulativeFitness: number;
  mutationThreshold: number;
  trend: GenotypeTrend;
  updatedAt: string;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  spawn: 'spawn',
  success: 'success',
  failure: 'failure',
  mutation: 'mutation',
  dispatch: 'dispatch',
  telemetry: 'telemetry',
} as const;

export type ActivitySeverity = typeof ActivitySeverity[keyof typeof ActivitySeverity];


export const ActivitySeverity = {
  info: 'info',
  success: 'success',
  warning: 'warning',
  danger: 'danger',
} as const;

export interface Activity {
  id: string;
  type: ActivityType;
  message: string;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  objectiveId?: string | null;
  timestamp: string;
  severity: ActivitySeverity;
}

export type SandboxJobStatus = typeof SandboxJobStatus[keyof typeof SandboxJobStatus];


export const SandboxJobStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export type SandboxJobTelemetry = {
  cpu: number;
  memory: number;
  network: number;
};

export interface SandboxJob {
  id: string;
  agentId: string;
  objectiveId: string;
  status: SandboxJobStatus;
  command: string;
  runtimeMs: number;
  /** @nullable */
  exitCode?: number | null;
  telemetry: SandboxJobTelemetry;
  startedAt: string;
}

export interface SandboxInput {
  /** @minLength 1 */
  agentId: string;
  /** @minLength 1 */
  objectiveId: string;
  /** @minLength 1 */
  command: string;
}

export type ListActivityParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};