# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivitySeverity } from './activitySeverity';
import type { ActivityType } from './activityType';

export interface Activity {
  id: string;
  type: ActivityType;
  message: string;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  objectiveId?: string | null;
  timestamp: Date;
  severity: ActivitySeverity;
}