# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentStatus } from './agentStatus';

export interface Agent {
  id: string;
  name: string;
  status: AgentStatus;
  strategy: string;
  frame: string;
  fitness: number;
  objectiveId: string;
  /** @nullable */
  parentId?: string | null;
  generation: number;
  runtimeMs: number;
  state: string;
  /** @nullable */
  failureSignature?: string | null;
  startedAt: Date;
}