# lib/api-zod/src/generated/types/genotype.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenotypeTrend } from './genotypeTrend';

export interface Genotype {
  id: string;
  frame: string;
  label: string;
  strategy: string;
  adjustedScore: number;
  fertilityScore: number;
  totalSpawns: number;
  successRate: number;
  cumulativeFitness: number;
  mutationThreshold: number;
  trend: GenotypeTrend;
  updatedAt: Date;
}