# lib/api-zod/src/generated/types/objective.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ObjectivePriority } from './objectivePriority';
import type { ObjectiveStatus } from './objectiveStatus';

export interface Objective {
  id: string;
  title: string;
  description: string;
  status: ObjectiveStatus;
  priority: ObjectivePriority;
  progress: number;
  agentCount: number;
  /** @nullable */
  rootAgentId?: string | null;
  createdAt: Date;
  updatedAt: Date;
}