# lib/api-zod/src/generated/types/objectiveInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ObjectiveInputPriority } from './objectiveInputPriority';

export interface ObjectiveInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  priority: ObjectiveInputPriority;
}