# lib/api-zod/src/generated/types/sandboxInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SandboxInput {
  /** @minLength 1 */
  agentId: string;
  /** @minLength 1 */
  objectiveId: string;
  /** @minLength 1 */
  command: string;
}