# lib/api-zod/src/generated/types/sandboxJob.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SandboxJobStatus } from './sandboxJobStatus';
import type { SandboxJobTelemetry } from './sandboxJobTelemetry';

export interface SandboxJob {
  id: string;
  agentId: string;
  objectiveId: string;
  status: SandboxJobStatus;
  command: string;
  runtimeMs: number;
  /** @nullable */
  exitCode?: number | null;
  telemetry: SandboxJobTelemetry;
  startedAt: Date;
}